confirm deletion of planner items
test plan: ensure the browser prompts you before deleting a
 planner item on the dashboard

fixes ADMIN-141

Reviewed-on: https://gerrit.instructure.com/129840
Tested-by: Jenkins
QA-Review: Dan Sasaki <dsasaki@instructure.com>
Product-Review: Jeremy Stanley <jeremy@instructure.com>

+++ src/components/UpdateItemTray/index.js
@@ -110,7 +110,9 @@ export class UpdateItemTray extends Component {
   }
 
   handleDeleteClick = () => {
-    this.props.onDeletePlannerItem(this.props.noteItem);
+    if (confirm(formatMessage('Are you sure you want to delete this planner item?'))) {
+      this.props.onDeletePlannerItem(this.props.noteItem);
+    }
   }
 
   findCurrentValue (field) {
